refactor(api): relocate Allegro config DTO + extract validator registry
Tech-review follow-up to the previous commit.

1. Move `AllegroConnectionConfigDto` from `apps/api/src/integrations/http/dto/`
   to `apps/api/src/integrations/application/dto/`. The DTO is now consumed
   by `ConnectionService.update()` (application layer) for server-side
   re-validation; importing it from the interface layer would have inverted
   the documented `interfaces → application → domain` dependency direction
   (`docs/architecture-overview.md`). The DTO was previously only used by
   future-controller-hookup code, so the move has zero runtime consumers
   to migrate; the FE doc-comment reference is updated to match.

2. Extract per-platform config validation into a small registry
   (`util/connection-config-validators.ts`) keyed by `platformType`. The
   previous `if (existing.platformType === 'allegro')` branch in
   `ConnectionService.update()` would have grown a switch-statement as
   PrestaShop / Shopify / etc. add their own config schemas; the registry
   makes new platforms plug in by registration.

Quality gate green: lint 0 errors, type-check passes, all 2084 tests pass.
