+++ apps/api/src/integrations/application/services/connection.service.ts
@@ -12,13 +12,10 @@
  */
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { randomUUID } from 'node:crypto';
-import { plainToInstance } from 'class-transformer';
-import { validate } from 'class-validator';
 import { IConnectionService } from '../interfaces/connection.service.interface';
 import { ConnectionCreateInput } from '../interfaces/connection.service.types';
 import { validateCredentialsShape } from '../credentials/credential-shape.validator';
-import { AllegroConnectionConfigDto } from '../../http/dto/allegro-connection-config.dto';
-import { flattenValidationErrors } from './util/flatten-validation-errors';
+import { CONNECTION_CONFIG_VALIDATORS } from './util/connection-config-validators';
 import {
   ConnectionPort,
   Connection,
@@ -266,14 +263,16 @@ export class ConnectionService implements IConnectionService {
 
       // #437 — close the DTO bypass on `Connection.config`. The HTTP-layer
       // `UpdateConnectionDto.config: Record<string, unknown>` erases the typed
-      // shape at the controller boundary, so the nested `AllegroConnectionConfigDto`
-      // decorators (location.countryCode, responsibleProducerId, safetyInformation
-      // discriminator, etc.) never run. Re-validate the platform-specific shape
-      // here, before persistence, so partial `sellerDefaults` blobs (the cause
-      // of the 2026-04-29 sandbox repro) are rejected at save time instead of
-      // surfacing as Allegro 422s at offer-create time.
-      if (patch.config !== undefined && existing.platformType === 'allegro') {
-        await this.validateAllegroConfig(patch.config);
+      // shape at the controller boundary, so the nested platform-specific
+      // decorators never run. Re-validate the platform-specific shape here,
+      // before persistence, so partial blobs (the cause of the 2026-04-29
+      // sandbox repro) are rejected at save time instead of surfacing as
+      // adapter 422s downstream. Per-platform validators are registered in
+      // `CONNECTION_CONFIG_VALIDATORS`; absence is a deliberate skip (no
+      // platform-specific shape to enforce yet).
+      const configValidator = CONNECTION_CONFIG_VALIDATORS[existing.platformType];
+      if (patch.config !== undefined && configValidator) {
+        await configValidator(patch.config);
       }
 
       const connection = await this.connectionPort.update(connectionId, patch);
@@ -322,23 +321,5 @@ export class ConnectionService implements IConnectionService {
     }
   }
 
-  private async validateAllegroConfig(config: Record<string, unknown>): Promise<void> {
-    const instance = plainToInstance(AllegroConnectionConfigDto, config);
-    // `whitelist: false` because the persisted `config` may carry adjacent keys
-    // (e.g. `adapterKey`, future per-platform tunables) that aren't part of
-    // `AllegroConnectionConfigDto` — we want shape-correctness on what the DTO
-    // *does* describe, not exhaustive ownership of the JSONB blob.
-    const errors = await validate(instance, {
-      whitelist: false,
-      forbidNonWhitelisted: false,
-    });
-    if (errors.length > 0) {
-      const issues = flattenValidationErrors(errors);
-      throw new BadRequestException({
-        message: 'Invalid Allegro connection config',
-        errors: issues,
-      });
-    }
-  }
 }
